% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals_functions_fnn.R
\name{.extractNetworkSpace}
\alias{.extractNetworkSpace}
\title{Extract Activation Spaces (Logits or Softmax) From Keras Model}
\usage{
.extractNetworkSpace(object, on = "logits")
}
\description{
Extract Activation Spaces (Logits or Softmax) From Keras Model
}
\keyword{internal}
