% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals_functions_fnn.R
\name{.resolveFNNPaths}
\alias{.resolveFNNPaths}
\title{Resolve package-relative paths in an FNN object (internal)}
\usage{
.resolveFNNPaths(object)
}
\description{
Resolve package-relative paths in an FNN object (internal)
}
\keyword{internal}
