% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals_functions_fnn.R
\name{.validateLayerOptions}
\alias{.validateLayerOptions}
\title{Validate layer option lists for FNN}
\usage{
.validateLayerOptions(layerList, requiredFields, layerName)
}
\arguments{
\item{layerList}{A named list of layer parameters.}

\item{requiredFields}{Character vector of required field names.}

\item{layerName}{Character. Name of the layer, used in error messages.}
}
\description{
Checks that a layer specification list contains all required fields with non-null values.
}
\keyword{internal}
