% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fplsda.R
\name{fit.fPLS_DA}
\alias{fit.fPLS_DA}
\title{Fit Partiel Least Square Regression and Discriminant Analysis on functionnal data (fPLS_DA)}
\usage{
fit.fPLS_DA(
  data,
  method = c("bsplines", "wavelets"),
  nbasis = 1050,
  ncomp = NULL,
  argvals = 2001:22664,
  rangeval = c(1999, 22664),
  n_folds = 5,
  filter = "la14",
  boundary = "periodic",
  n.levels = NULL,
  level = 3
)
}
\arguments{
\item{data}{Matrice de donnees MSP transformees pour l'entrainement du modele.}

\item{method}{Méthode de décomposition en bases fonctionnelles : "bsplines" ou "wavelets".}

\item{nbasis}{an integer variable specifying the number of basis functions (1050 par défaut pour bsplines).}

\item{ncomp}{Nombre de composantes PLS à retenir (par défaut 20).}

\item{argvals}{description}

\item{rangeval}{description}

\item{n_folds}{Nombre de plis pour la validation croisée (par défaut 5).}

\item{filter}{Either a \code{wt.filter} object, a character string
    indicating which wavelet filter to use in the decomposition, or a
    numeric vector of wavelet coefficients (not scaling coefficients). See
    \code{help(wt.filter)} for acceptable filter names.}

\item{boundary}{A character string indicating which boundary method to
    use. \code{boundary = "periodic"} and \code{boundary = "reflection"}
    are the only supported methods at this time.}

\item{n.levels}{An integer specifying the level of the decomposition.
    By default this is the value J such that the length of \eqn{X} is at
    least as great as the length of the level \eqn{J} wavelet filter,
    but less than the length of the level \eqn{J+1} wavelet
    filter. Thus, \eqn{J \le \log{(\frac{N-1}{L-1}+1)}}{j <=
      log((N-1)/(L-1)+1)}, where \eqn{N} is the length of \eqn{X}.}
}
\description{
Fit Partiel Least Square Regression and Discriminant Analysis on functionnal data (fPLS_DA)
}
