% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_maldi}
\alias{process_maldi}
\title{Importation et Pretraitement des replicas biologiques de souches}
\usage{
process_maldi(data_file, method_baseline = "SNIP", halfWindowSize = 10)
}
\arguments{
\item{data_file}{Chemin vers le dossier contenant les sous-dossiers de souches.}

\item{method_baseline}{Methode de retrait de la ligne de base ("SNIP" par defaut).}

\item{halfWindowSize}{Fenetre de valeurs pour le lissage SavitzkyGolay.}
}
\value{
Une matrice d'intensites (MSP) ou chaque ligne est un replica biologique normalise.
}
\description{
Importation et Pretraitement des replicas biologiques de souches
}
