% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals_functions.R
\name{.fdaDecomposition}
\alias{.fdaDecomposition}
\title{Functional decomposition of spectral data (internal)}
\usage{
.fdaDecomposition(
  data,
  method,
  nbasis,
  rangeval,
  argvals,
  filter,
  boundary,
  nlevels,
  level
)
}
\arguments{
\item{data}{A numeric matrix of spectral intensities (rows = samples, columns = m/z values).}

\item{method}{Character. Either \code{"bsplines"} or \code{"wavelets"}.}

\item{nbasis}{Integer. Number of B-spline basis functions. Used only when  \code{method = "bsplines"}.}

\item{rangeval}{Numeric vector of length 2. Domain range for the B-spline basis. Used only when \code{method = "bsplines"}.}

\item{argvals}{Numeric vector. Evaluation grid (m/z indices) for B-spline projection. Used only when \code{method = "bsplines"}.}

\item{filter}{Character. The wavelet filter type passed to \code{\link[wavelets]{dwt}} when \code{method = "wavelets"}.}

\item{boundary}{Character. The boundary handling method passed to \code{\link[wavelets]{dwt}}. Default is \code{"periodic"}.}

\item{nlevels}{Integer or \code{NULL}. Total number of wavelet decomposition levels. If \code{NULL}, set to the maximum possible level. Used only when \code{method = "wavelets"}.}

\item{level}{Integer. Wavelet resolution level from which coefficients are extracted. Used only when \code{method = "wavelets"}.}
}
\value{
A named list with the following elements:
\item{coef_matrix}{Numeric matrix of decomposition coefficients (rows = samples).}
\item{basis_obj}{The B-spline basis object (\code{\link[fda:create.bspline.basis]{fda::create.bspline.basis()}}), or \code{NULL} if \code{method = "wavelets"}.}
\item{wt_output}{A list of raw \code{\link[wavelets:dwt]{wavelets::dwt()}} objects (one per sample), or \code{NULL} if \code{method = "bsplines"}.}
\item{nlevels}{The effective number of wavelet decomposition levels used or \code{NULL} if \code{method = "bsplines"}.}
\item{level}{The effective wavelet resolution level used, or \code{NULL} if \code{method = "bsplines"}.}
}
\description{
Projects a raw spectral matrix into a reduced coefficient matrix using either a B-spline basis or a Discrete Wavelet Transform (DWT).
}
\keyword{internal}
