% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_spectra.R
\name{plot_functions}
\alias{plot_functions}
\alias{plot_spectra_3d}
\alias{plot_spectra_2d}
\alias{plot_pls_var}
\alias{plot_pls_weights}
\alias{plot_spectra_umap}
\title{Visualisations des resultats de l'analyse discriminante
PROJECTION 3D dans l'espace des composantes de la PLS
PROJECTION 2D dans le plan des composantes de la PLS
Visualisation graphique de la VARIANCE EXPLIQUEE des composantes PLS (Scree Plot)}
\usage{
plot_spectra_3d(object, comp = c(1, 2, 3))

plot_spectra_2d(object, comp = c(1, 2), show_proba = FALSE)

plot_pls_var(object)

plot_pls_weights(object, comp = 1)

plot_spectra_umap(object, n_neighbors = 15, min_dist = 0.1)
}
\arguments{
\item{object}{a fitted object of class inheriting from "fPLS_DA".}

\item{comp}{Axes de projection des scores (par défaut 1 et 2)}

\item{show_proba}{Booleen. Si TRUE, colorie par groupe et ajuste la transparence (alpha) selon la probabilite de prediction.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Visualisations des resultats de l'analyse discriminante
PROJECTION 3D dans l'espace des composantes de la PLS
PROJECTION 2D dans le plan des composantes de la PLS
Visualisation graphique de la VARIANCE EXPLIQUEE des composantes PLS (Scree Plot)
}
