% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fplsda.R
\name{predict.fPLS_DA}
\alias{predict.fPLS_DA}
\title{Predict Method for fPLS_DA fits}
\usage{
\method{predict}{fPLS_DA}(object, newdata, threshold = NULL)
}
\arguments{
\item{object}{a fitted object of class inheriting from "fPLS_DA".}

\item{newdata}{an optionally data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.   ###################" Revoir le newdata et le rendre optionnel}

\item{threshold}{Seuil de validation du type de la souche}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predict Method for fPLS_DA fits
}
